Added realisation backends (string, simplenlg).
